+++ dear-diary/src/App.tsx
@@ -4,21 +4,7 @@ import HomePage from './containers/HomePage/HomePage';
 import 'react-toastify/dist/ReactToastify.css';
 import { useAppSelector } from './redux/hooks';
 import { ReactElement } from 'react';
-
-const routes = [
-  { path: '/', element: <SignInPage /> },
-  { path: '/home', element: <HomePage /> },
-];
-
-const PrivateRoute = ( element: ReactElement, path: string ) => {
-  const isAuthenticated = useAppSelector( (state) => state.user.isLogged );
-  const defaultElement = <SignInPage/>;
-
-  return ( isAuthenticated ? 
-  < Route path={path} element={element} /> : 
-  < Route path={path} element={defaultElement} />)
-  
-};
+import { routes, PrivateRoute } from './routes';
 
 function App() {
   return (
