+++ dear-diary/src/utilities/validation.ts
@@ -5,7 +5,6 @@ function validateDiaryCard(title: string, description: string): boolean {
   if (!title.trim() || !description.trim()) {
     return false;
   }
-
   return true;
 }
